* All added NSTokenView initial version

+++ Source/Bodo/ImageRepository.cs
@@ -37,7 +37,7 @@ namespace Bodo
 
 		public int ImageCount
 		{
-			get { return (int)_fetchResult.Count; }
+			get { return _fetchResult.Count; }
 		}
 	}
 }
